requirements.txt
flask
pandas
scikit-learn
joblib
PyPDF2
python-docx
psycopg2-binary